+++ tests/job_tests.cpp
@@ -1,5 +1,5 @@
 //* job_tests
-#ifdef ENABLE_TESTS
+#ifdef FRESA_ENABLE_TESTS
 
 #include "unit_test.h"
 #include "jobs.h"