+++ tests/math_tests.cpp
@@ -1,5 +1,5 @@
 //* types_tests
-#ifdef ENABLE_TESTS
+#ifdef FRESA_ENABLE_TESTS
 
 #include "unit_test.h"
 #include "fresa_math.h"